+++ .github/workflows/test.yaml
@@ -1,0 +1,35 @@
+name: tests
+on:
+  push:
+    paths-ignore:
+      - '**/README.md'
+      - 'static/**/*'
+  pull_request:
+    paths-ignore:
+      - '**/README.md'
+      - 'static/**/*'
+jobs:
+  test-core-envs:
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: [3.8]
+        poetry-version: [1.3]
+        os: [ubuntu-22.04, macos-latest, windows-latest]
+    runs-on: ${{ matrix.os }}
+    environment: ci
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Run image
+        uses: abatilo/actions-poetry@v2.0.0
+        with:
+          poetry-version: ${{ matrix.poetry-version }}
+
+      # classic control tests
+      - name: Install core dependencies
+        run: poetry install --with dev
+      - name: Run core tests
+        run: poetry run pytest tests/test_e2e_ci.py
